learning_materials/lang_chain/pyproject.toml
[project]
name = "lang-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[anthropic,google-genai,mistralai,openai]>=1.1.0",
    "python-dotenv>=1.2.1",
]
